Multicore compiling
Added -j{nproc} after every make command, could be improved by multipling nproc with * 1.5 because `The 1.5 number is because of the noted I/O bound problem. It is a rule of thumb. About 1/3 of the jobs will be waiting for I/O, so the remaining jobs will be using the available cores. A number greater than the cores is better and you could even go as high as 2x.`